Export gatt module and process function for individual invocation
If you're using noble or something like that for recieving BLE advertisements, it's handy to be able to invoke just parts of the processing. That's already the case for most of the gap module, but the gatt module was buried. This makes it a public property so it can be invoked too!